Build/Test Tools: Add matrix test for databases
The suitability of databases needs to be tested with CI.

Add testing for MySQL and MariaDB on the their earliest supported
version, and the latest version.

MariaDB Testing is also done on their non-released version by using
a container image produced from their CI. This image contains reviewed
and tested work of the MariaDB developers. This is a good opporunity
to catch pre-release regressions on their part.

This is based off the end to end tests, but now we are applying them
across database versions.

Docker debug information uses \s to show the client and server
information. The MariaDB Testing images also contain /manifest.txt
containing the git version of build to report regressions.

References:
* [https://mariadb.org/new-service-quay-io-mariadb-foundation-mariadb-devel/ MariaDB Blog of quay.io service]

Trac ticket: